refactor: create home component
